fix install issue

+++ terraform/main.tf
@@ -100,28 +100,50 @@ resource "aws_instance" "py_server" {
   instance_type = "t2.micro"
   key_name      = aws_key_pair.deployer.key_name
   user_data     = <<-EOF
-              #!/bin/bash
+                    #!/bin/bash
 
-              PROJ=tfgha
-              POETRY=/home/ubuntu/.local/bin
-              WORKDIR=/home/ubuntu
+                    PROJ="tfgha"
+                    POETRY="/home/ubuntu/.local/bin"
+                    WORKDIR="/home/ubuntu"
 
-              sudo apt-get update
-              sudo apt-get install -y python3 python3-pip git curl
-              cd $WORKDIR
-              git clone https://github.com/proquickly/$PROJ.git
-              sudo chown -R ubuntu:ubuntu $WORKDIR
-              su - ubuntu -c '
-                python3 -m pip install -U poetry
-                cd $WORKDIR/$PROJ
-                export PATH="$WORKDIR/.local/bin:$PATH"
-                rm poetry.lock
-                $POETRY/poetry install
-                cd $WORKDIR/$PROJ/src/$PROJ
+                    # Update package lists
+                    if ! sudo apt-get update; then
+                      echo "Failed to update package lists"
+                      exit 1
+                    fi
 
-                nohup $POETRY/poetry run python app.py &
-              '
-              EOF
+                    # Install necessary packages
+                    if ! sudo apt-get install -y python3 python3-pip git curl; then
+                      echo "Failed to install packages"
+                      exit 1
+                    fi
+
+                    # Change to the working directory
+                    cd "$WORKDIR" || { echo "Failed to change directory to $WORKDIR"; exit 1; }
+
+                    # Clone the project
+                    if ! git clone "https://github.com/proquickly/$PROJ.git"; then
+                      echo "Failed to clone the repository"
+                      exit 1
+                    fi
+
+                    # Change ownership
+                    if ! sudo chown -R ubuntu:ubuntu "$WORKDIR"; then
+                      echo "Failed to change ownership"
+                      exit 1
+                    fi
+
+                    # Run as the ubuntu user
+                    sudo -u ubuntu bash <<EOF
+                      set -e
+                      python3 -m pip install -U poetry
+                      cd "$WORKDIR/$PROJ"
+                      export PATH="$WORKDIR/.local/bin:\$PATH"
+                      [ -f poetry.lock ] && rm poetry.lock
+                      $POETRY/poetry install
+                      cd "$WORKDIR/$PROJ/src/$PROJ"
+                      nohup $POETRY/poetry run python app.py &
+                    EOF
 
   tags = {
     Name = "FlaskAppInstance"
